requirements.txt
pytest
adlfs==2025.8.0
aiobotocore==3.1.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiooss2==0.2.11
aiosignal==1.4.0
alembic==1.18.3
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
argcomplete==3.6.3
async-timeout==5.0.1
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
azure-core==1.38.0
azure-datalake-store==0.0.53
azure-identity==1.25.1
azure-storage-blob==12.28.0
bcrypt==5.0.0
billiard==4.2.4
bleach==6.3.0
blinker==1.9.0
botocore==1.42.42
cachetools==6.2.6
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
databricks-sdk==0.85.0
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==1.0.0
dvc==3.66.1
dvc-azure==3.1.0
dvc-data==3.18.2
dvc-gdrive==3.0.1
dvc-gs==3.0.2
dvc-hdfs==3.0.0
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-oss==3.0.0
dvc-render==1.0.2
dvc-s3==3.3.0
dvc-ssh==4.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dvc-webdav==3.0.1
dvc-webhdfs==3.1.0
entrypoints==0.4
exceptiongroup==1.3.1
fastapi==0.128.4
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
funcy==2.0
gcsfs==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-cloud-storage-control==1.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.1
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
gto==1.9.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huey==2.6.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.3
kaggle==1.7.4.5
kiwisolver==1.4.9
knack==0.13.0
kombu==5.6.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.1
networkx==3.4.2
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.13.0.92
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
oss2==2.18.4
ossfs==2025.5.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.1.0
platformdirs==4.5.1
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
PyDrive2==1.21.3
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.11.0
pyOpenSSL==24.2.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rsa==4.9.1
ruamel.yaml==0.19.1
s3fs==2026.2.0
scikit-learn==1.7.2
scipy==1.15.3
scmrepo==3.6.1
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skops==0.13.0
smmap==5.0.2
SQLAlchemy==2.0.46
sqlparse==0.5.5
sqltrie==0.11.2
sshfs==2025.11.0
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.4.0
tomlkit==0.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
voluptuous==0.16.0
waitress==3.0.2
wcwidth==0.6.0
webdav4==0.10.0
webencodings==0.5.1
Werkzeug==3.1.5
wrapt==2.1.1
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
